Make clipboard configurable
In some cases the primary or secondary clipboard is more interesting to
use than "clipboard" is. Let's make this thing configurable. This patch
centralizes the choice for clipboard into a variable CLIPBOARD, which is
then used for all xclip calls. Ideally there should be a command line
option to choose the specific clipboard to use for this call.